requirements.txt
fastapi
aiologger[aiofiles]
pip-chill
uvicorn
gunicorn
python-dotenv
duckduckgo_search
openai
requests
